+++ stockfish-core/src/core/rank.rs
@@ -32,6 +32,28 @@ impl Rank {
     }
 }
 
+impl IntoIterator for Rank {
+    type Item     = Square;
+    type IntoIter = std::array::IntoIter<Square, 8>;
+
+    #[inline]
+    #[must_use]
+    fn into_iter(self) -> Self::IntoIter {
+        use Square as S;
+
+        match self {
+            Self::_1 => [S::A1, S::B1, S::C1, S::D1, S::E1, S::F1, S::G1, S::H1],
+            Self::_2 => [S::A2, S::B2, S::C2, S::D2, S::E2, S::F2, S::G2, S::H2],
+            Self::_3 => [S::A3, S::B3, S::C3, S::D3, S::E3, S::F3, S::G3, S::H3],
+            Self::_4 => [S::A4, S::B4, S::C4, S::D4, S::E4, S::F4, S::G4, S::H4],
+            Self::_5 => [S::A5, S::B5, S::C5, S::D5, S::E5, S::F5, S::G5, S::H5],
+            Self::_6 => [S::A6, S::B6, S::C6, S::D6, S::E6, S::F6, S::G6, S::H6],
+            Self::_7 => [S::A7, S::B7, S::C7, S::D7, S::E7, S::F7, S::G7, S::H7],
+            Self::_8 => [S::A8, S::B8, S::C8, S::D8, S::E8, S::F8, S::G8, S::H8],
+        }.into_iter()
+    }
+}
+
 impl const From<Rank> for char {
     #[inline]
     fn from(value: Rank) -> Self {
@@ -46,26 +68,6 @@ impl const From<Square> for Rank {
             Self::from_u8(s.rank_index()).unwrap(),
             Self::from_u8_unchecked(s.rank_index()),
         )
-    }
-}
-
-impl const IntoIterator for Rank {
-    type Item     = Square;
-    type IntoIter = std::ops::RangeInclusive<Square>;
-
-    #[inline]
-    #[must_use]
-    fn into_iter(self) -> Self::IntoIter {
-        match self {
-            Self::_1 => Square::A1 ..= Square::H1,
-            Self::_2 => Square::A2 ..= Square::H2,
-            Self::_3 => Square::A3 ..= Square::H3,
-            Self::_4 => Square::A4 ..= Square::H4,
-            Self::_5 => Square::A5 ..= Square::H5,
-            Self::_6 => Square::A6 ..= Square::H6,
-            Self::_7 => Square::A7 ..= Square::H7,
-            Self::_8 => Square::A8 ..= Square::H8,
-        }
     }
 }
 
@@ -120,6 +122,14 @@ mod tests {
     fn rank_bitor_file() {
         for square in Square::iter() {
             assert_eq!(square.rank() | square.file(), square);
+        }
+    }
+
+    #[test]
+    fn rank_into_iter() {
+        for rank in Rank::iter() {
+            assert_eq!(8, rank.into_iter().count());
+            assert!(rank.into_iter().is_sorted_by_key(Square::rank));
         }
     }
 }