+++ stockfish-core/CHANGELOG.md
@@ -12,6 +12,12 @@ adheres to [Semantic Versioning][semver].
 - `File` and `Rank` now implement `BitOr` against themselves
 - `Bitboard` can now `BitAnd`, `BitOr`, and `BitXor` against `File` and `Rank`
 
+### Changed
+
+- `File` and `Rank` now *both* allow iteration over the squares they contain,
+  but the type of their `IntoIterator` is changed to
+  `std::array::IntoIter<Square, 8>`
+
 ## [0.1.1][https://github.com/stouset/stockfish-rs/tree/stockfish-core-v0.1.1] - 2023-01-13
 
 ### Fixed